refactor(k8s): frame + split the detector layer (was problems.go)
problems.go was the v0 all-in-one "problems" feature; in the issues architecture
its real role is the DETECTOR LAYER — it reads the cache and emits []Detection
(raw findings) that internal/issues classifies + groups, and that the dashboard +
MCP also consume. The name and 1024-line god-file structure hid that.

- Renamed problems.go → detection.go (matches the Detection type; P1 follow-through).
- Split by detector to match the clean single-purpose files in internal/issues:
  detection.go (Detection type + the core DetectProblems workload/pod/service/etc
  detector + shared helpers), capi.go (DetectCAPIProblems), gitops.go
  (DetectGitOpsProblems + Argo/Flux helpers). Siblings health.go/missing_refs.go/
  scheduling.go were already separate detectors.
- Added a layer-framing doc on Detection: detectors → classify/group (issues) →
  render; this is the bottom tier, not a parallel surface to issues.
