+++ internal/k8s/detection.go
@@ -2,22 +2,32 @@ package k8s
 
 import (
 	"fmt"
-	"log"
 	"sort"
 	"strings"
 	"time"
 
-	"github.com/skyhook-io/radar/pkg/conditions"
 	appsv1 "k8s.io/api/apps/v1"
 	autoscalingv2 "k8s.io/api/autoscaling/v2"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/util/intstr"
 )
 
-// Detection is a transport-neutral cluster issue.
+// Detection is a transport-neutral raw operational finding emitted by the
+// detector layer — a failing Deployment, a crashlooping pod, a dangling
+// reference, a degraded Argo app. It carries NO classification, grouping, or
+// ranking; those are the issues layer's job.
+//
+// detection.go and its sibling detectors (health.go, missing_refs.go,
+// scheduling.go, capi.go, gitops.go) ARE that detector layer: each reads the
+// live cache and returns []Detection. internal/issues is the layer ABOVE them —
+// it classifies each Detection into a symptom Category, resolves its Subject,
+// and folds replica fan-out into the public grouped Issue model; the home
+// dashboard and MCP also consume Detections directly. So this is the bottom
+// tier of the pipeline (detectors → classify/group → render) — the generalized
+// successor to the v0 standalone "problems" feature, NOT a parallel surface to
+// issues.
 type Detection struct {
 	Kind            string
 	Namespace       string
@@ -618,407 +628,3 @@ func failedJobCondition(job *batchv1.Job) *batchv1.JobCondition {
 // Checks both status.phase and the rich condition system (Ready, InfrastructureReady,
 // ControlPlaneReady, BootstrapReady, NodeHealthy, TopologyReconciled).
 // Returns nil if CAPI is not installed in the cluster.
-func DetectCAPIProblems(dynamicCache *DynamicResourceCache, discovery *ResourceDiscovery, namespace string) []Detection {
-	if dynamicCache == nil || discovery == nil {
-		return nil
-	}
-
-	var problems []Detection
-	now := time.Now()
-
-	// Helper: list CAPI resources by kind
-	listCAPI := func(kind, group string) []*unstructured.Unstructured {
-		if group != "" {
-			gvr, ok := discovery.GetGVRWithGroup(kind, group)
-			if !ok {
-				return nil // CRD not installed — expected
-			}
-			resources, err := dynamicCache.List(gvr, namespace)
-			if err != nil {
-				log.Printf("[capi-problems] Failed to list %s (%s): %v", kind, group, err)
-				return nil
-			}
-			return resources
-		}
-		gvr, ok := discovery.GetGVR(kind)
-		if !ok {
-			return nil // CRD not installed — expected
-		}
-		resources, err := dynamicCache.List(gvr, namespace)
-		if err != nil {
-			log.Printf("[capi-problems] Failed to list %s: %v", kind, err)
-			return nil
-		}
-		return resources
-	}
-
-	// Shared condition reader: conditions.FindFalseCondition (one source of truth
-	// across the CAPI/GitOps detectors + the issues generic fallback).
-
-	const capiGroup = "cluster.x-k8s.io"
-	const capiCPGroup = "controlplane.cluster.x-k8s.io"
-
-	// -----------------------------------------------------------------------
-	// CAPI Cluster problems
-	// -----------------------------------------------------------------------
-	for _, cl := range listCAPI("Cluster", capiGroup) {
-		ageDur := now.Sub(cl.GetCreationTimestamp().Time)
-
-		// Phase-based: Failed
-		phase, _, _ := unstructured.NestedString(cl.Object, "status", "phase")
-		if strings.EqualFold(phase, "failed") {
-			problems = append(problems, Detection{
-				Kind: "Cluster", Namespace: cl.GetNamespace(), Name: cl.GetName(), Group: capiGroup,
-				Severity: "critical", Reason: "Cluster in Failed phase",
-				Age: FormatAge(ageDur), AgeSeconds: int64(ageDur.Seconds()),
-				Duration: FormatAge(ageDur), DurationSeconds: int64(ageDur.Seconds()),
-			})
-			continue // don't double-report conditions
-		}
-
-		// Condition-based: InfrastructureReady, ControlPlaneReady, Ready, TopologyReconciled
-		if ct, reason, msg, dur, ok := conditions.FindFalseCondition(cl,
-			"Ready", "InfrastructureReady", "ControlPlaneReady", "TopologyReconciled",
-		); ok {
-			severity := "high"
-			if ct == "InfrastructureReady" || ct == "ControlPlaneReady" {
-				severity = "critical"
-			}
-			displayReason := reason
-			if displayReason == "" {
-				displayReason = ct + "=False"
-			}
-			d := dur
-			if d == 0 {
-				d = ageDur
-			}
-			problems = append(problems, Detection{
-				Kind: "Cluster", Namespace: cl.GetNamespace(), Name: cl.GetName(), Group: capiGroup,
-				Severity: severity, Reason: displayReason, Message: msg,
-				Age: FormatAge(ageDur), AgeSeconds: int64(ageDur.Seconds()),
-				Duration: FormatAge(d), DurationSeconds: int64(d.Seconds()),
-			})
-		}
-	}
-
-	// -----------------------------------------------------------------------
-	// CAPI Machine problems
-	// -----------------------------------------------------------------------
-	for _, m := range listCAPI("Machine", "cluster.x-k8s.io") {
-		ageDur := now.Sub(m.GetCreationTimestamp().Time)
-		phase, _, _ := unstructured.NestedString(m.Object, "status", "phase")
-
-		// Phase-based: Failed
-		if strings.EqualFold(phase, "failed") {
-			// Include the condition message for richer context
-			_, _, msg, _, _ := conditions.FindFalseCondition(m, "Ready", "InfrastructureReady", "BootstrapReady")
-			problems = append(problems, Detection{
-				Kind: "Machine", Namespace: m.GetNamespace(), Name: m.GetName(), Group: capiGroup,
-				Severity: "critical", Reason: "Machine in Failed phase", Message: msg,
-				Age: FormatAge(ageDur), AgeSeconds: int64(ageDur.Seconds()),
-				Duration: FormatAge(ageDur), DurationSeconds: int64(ageDur.Seconds()),
-			})
-			continue
-		}
-
-		// Phase-based: stuck Provisioning > 10m
-		if strings.EqualFold(phase, "provisioning") && ageDur > 10*time.Minute {
-			_, reason, msg, _, _ := conditions.FindFalseCondition(m, "InfrastructureReady", "BootstrapReady")
-			displayReason := fmt.Sprintf("Stuck provisioning for %s", FormatAge(ageDur))
-			if reason != "" {
-				displayReason += " (" + reason + ")"
-			}
-			problems = append(problems, Detection{
-				Kind: "Machine", Namespace: m.GetNamespace(), Name: m.GetName(), Group: capiGroup,
-				Severity: "high", Reason: displayReason, Message: msg,
-				Age: FormatAge(ageDur), AgeSeconds: int64(ageDur.Seconds()),
-				Duration: FormatAge(ageDur), DurationSeconds: int64(ageDur.Seconds()),
-			})
-			continue
-		}
-
-		// Condition-based: BootstrapReady=False, NodeHealthy=False, InfrastructureReady=False
-		// (catches problems that phase alone misses, e.g. Running phase but NodeHealthy=False)
-		if ct, reason, msg, dur, ok := conditions.FindFalseCondition(m,
-			"BootstrapReady", "NodeHealthy", "InfrastructureReady",
-		); ok {
-			severity := "high"
-			if ct == "BootstrapReady" {
-				severity = "critical"
-			}
-			displayReason := reason
-			if displayReason == "" {
-				displayReason = ct + "=False"
-			}
-			d := dur
-			if d == 0 {
-				d = ageDur
-			}
-			problems = append(problems, Detection{
-				Kind: "Machine", Namespace: m.GetNamespace(), Name: m.GetName(), Group: capiGroup,
-				Severity: severity, Reason: displayReason, Message: msg,
-				Age: FormatAge(ageDur), AgeSeconds: int64(ageDur.Seconds()),
-				Duration: FormatAge(d), DurationSeconds: int64(d.Seconds()),
-			})
-		}
-	}
-
-	// -----------------------------------------------------------------------
-	// CAPI MachineDeployment problems: ready < desired for > 5m
-	// -----------------------------------------------------------------------
-	for _, md := range listCAPI("MachineDeployment", "") {
-		desired, _, _ := unstructured.NestedInt64(md.Object, "spec", "replicas")
-		ready, _, _ := unstructured.NestedInt64(md.Object, "status", "readyReplicas")
-		if desired > 0 && ready < desired {
-			ageDur := now.Sub(md.GetCreationTimestamp().Time)
-			if ageDur > 5*time.Minute {
-				_, reason, msg, _, _ := conditions.FindFalseCondition(md, "Ready", "Available")
-				displayReason := fmt.Sprintf("%d/%d machines ready", ready, desired)
-				if reason != "" {
-					displayReason += " (" + reason + ")"
-				}
-				problems = append(problems, Detection{
-					Kind: "MachineDeployment", Namespace: md.GetNamespace(), Name: md.GetName(), Group: capiGroup,
-					Severity: "high", Reason: displayReason, Message: msg,
-					Age: FormatAge(ageDur), AgeSeconds: int64(ageDur.Seconds()),
-					Duration: FormatAge(ageDur), DurationSeconds: int64(ageDur.Seconds()),
-				})
-			}
-		}
-	}
-
-	// -----------------------------------------------------------------------
-	// CAPI KubeadmControlPlane problems: Ready=False or replicas mismatch
-	// -----------------------------------------------------------------------
-	for _, kcp := range listCAPI("KubeadmControlPlane", "") {
-		ageDur := now.Sub(kcp.GetCreationTimestamp().Time)
-		desired, _, _ := unstructured.NestedInt64(kcp.Object, "spec", "replicas")
-		ready, _, _ := unstructured.NestedInt64(kcp.Object, "status", "readyReplicas")
-
-		if ct, reason, msg, dur, ok := conditions.FindFalseCondition(kcp,
-			"Ready", "Available", "CertificatesAvailable", "MachinesReady",
-		); ok {
-			severity := "critical"
-			displayReason := reason
-			if displayReason == "" {
-				displayReason = ct + "=False"
-			}
-			if desired > 0 && ready < desired {
-				displayReason = fmt.Sprintf("%d/%d CP replicas ready, %s", ready, desired, displayReason)
-			}
-			d := dur
-			if d == 0 {
-				d = ageDur
-			}
-			problems = append(problems, Detection{
-				Kind: "KubeadmControlPlane", Namespace: kcp.GetNamespace(), Name: kcp.GetName(), Group: capiCPGroup,
-				Severity: severity, Reason: displayReason, Message: msg,
-				Age: FormatAge(ageDur), AgeSeconds: int64(ageDur.Seconds()),
-				Duration: FormatAge(d), DurationSeconds: int64(d.Seconds()),
-			})
-		}
-	}
-
-	// -----------------------------------------------------------------------
-	// CAPI MachineHealthCheck: actively remediating
-	// -----------------------------------------------------------------------
-	for _, mhc := range listCAPI("MachineHealthCheck", "") {
-		expected, _, _ := unstructured.NestedInt64(mhc.Object, "status", "expectedMachines")
-		healthy, _, _ := unstructured.NestedInt64(mhc.Object, "status", "currentHealthy")
-		if expected > 0 && healthy < expected {
-			ageDur := now.Sub(mhc.GetCreationTimestamp().Time)
-			problems = append(problems, Detection{
-				Kind: "MachineHealthCheck", Namespace: mhc.GetNamespace(), Name: mhc.GetName(), Group: capiGroup,
-				Severity:        "high",
-				Reason:          fmt.Sprintf("Remediating: %d/%d healthy", healthy, expected),
-				Age:             FormatAge(ageDur),
-				AgeSeconds:      int64(ageDur.Seconds()),
-				Duration:        FormatAge(ageDur),
-				DurationSeconds: int64(ageDur.Seconds()),
-			})
-		}
-	}
-
-	return problems
-}
-
-const (
-	argoGroup   = "argoproj.io"
-	fluxKustGrp = "kustomize.toolkit.fluxcd.io"
-	fluxHelmGrp = "helm.toolkit.fluxcd.io"
-)
-
-// DetectGitOpsProblems surfaces failing GitOps reconcilers — ArgoCD Applications
-// and Flux Kustomizations/HelmReleases — that the generic CRD-condition fallback
-// structurally misses. Argo encodes health and sync in dedicated status
-// sub-objects (status.health.status, status.sync.status) rather than as
-// status.conditions[type=Ready] entries, so FindFalseCondition never sees a
-// Degraded/Missing/OutOfSync app; and Argo "ComparisonError" lives only in
-// status.conditions[].type (no status=False). This detector reads each
-// controller's real shape. Wired like DetectCAPIProblems; detectGenericCRDIssues
-// skips exactly the kinds handled here (isCuratedCRDKind) so there is no
-// double-report, while leaving sibling kinds (e.g. Argo Rollout) to the generic
-// path.
-func DetectGitOpsProblems(dynamicCache *DynamicResourceCache, discovery *ResourceDiscovery, namespace string) []Detection {
-	if dynamicCache == nil || discovery == nil {
-		return nil
-	}
-	now := time.Now()
-	list := func(kind, group string) []*unstructured.Unstructured {
-		gvr, ok := discovery.GetGVRWithGroup(kind, group)
-		if !ok {
-			return nil // controller not installed — expected
-		}
-		items, err := dynamicCache.List(gvr, namespace)
-		if err != nil {
-			log.Printf("[gitops-problems] Failed to list %s.%s: %v", kind, group, err)
-			return nil
-		}
-		return items
-	}
-
-	var problems []Detection
-	problems = append(problems, detectArgoAppProblems(list("Application", argoGroup), now)...)
-	problems = append(problems, detectFluxProblems(list("Kustomization", fluxKustGrp), "Kustomization", fluxKustGrp, now)...)
-	problems = append(problems, detectFluxProblems(list("HelmRelease", fluxHelmGrp), "HelmRelease", fluxHelmGrp, now)...)
-	return problems
-}
-
-func gitopsProblem(kind, group, ns, name, severity, reason, message string, age time.Duration) Detection {
-	return Detection{
-		Kind:            kind,
-		Group:           group,
-		Namespace:       ns,
-		Name:            name,
-		Severity:        severity,
-		Reason:          reason,
-		Message:         message,
-		Age:             FormatAge(age),
-		AgeSeconds:      int64(age.Seconds()),
-		Duration:        FormatAge(age),
-		DurationSeconds: int64(age.Seconds()),
-	}
-}
-
-// detectArgoAppProblems reads ArgoCD Application health/sync. Precision gates,
-// all load-bearing (a manual or suspended app legitimately sits OutOfSync/Missing
-// and must NOT flag): skip Suspended/Progressing health and an in-flight sync
-// (operationState.phase=Running); flag Degraded regardless of policy (live
-// resources are unhealthy); flag Missing/OutOfSync only for auto-synced apps; and
-// always flag a ComparisonError/InvalidSpecError condition (the sync=Unknown
-// app-path-not-found case the generic path also can't see). One row per app,
-// most-specific cause first.
-func detectArgoAppProblems(apps []*unstructured.Unstructured, now time.Time) []Detection {
-	var out []Detection
-	for _, app := range apps {
-		ns, name := app.GetNamespace(), app.GetName()
-		age := now.Sub(app.GetCreationTimestamp().Time)
-		health, _, _ := unstructured.NestedString(app.Object, "status", "health", "status")
-		sync, _, _ := unstructured.NestedString(app.Object, "status", "sync", "status")
-		phase, _, _ := unstructured.NestedString(app.Object, "status", "operationState", "phase")
-
-		if strings.EqualFold(health, "Suspended") || strings.EqualFold(health, "Progressing") || strings.EqualFold(phase, "Running") {
-			continue
-		}
-
-		if ct, msg, ok := argoErrorCondition(app); ok {
-			out = append(out, gitopsProblem("Application", argoGroup, ns, name, "high", ct, msg, age))
-			continue
-		}
-		if strings.EqualFold(health, "Degraded") {
-			out = append(out, gitopsProblem("Application", argoGroup, ns, name, "critical",
-				"HealthDegraded", "Application health is Degraded (managed resources are unhealthy)", age))
-			continue
-		}
-		automated := argoIsAutomated(app)
-		if strings.EqualFold(health, "Missing") && automated {
-			out = append(out, gitopsProblem("Application", argoGroup, ns, name, "high",
-				"HealthMissing", "auto-synced Application's managed resources are missing from the cluster", age))
-			continue
-		}
-		if strings.EqualFold(sync, "OutOfSync") && automated {
-			out = append(out, gitopsProblem("Application", argoGroup, ns, name, "high",
-				"OutOfSync", "auto-synced Application has drifted from the desired manifests", age))
-		}
-	}
-	return out
-}
-
-// argoIsAutomated reports whether spec.syncPolicy.automated is present — i.e. the
-// app is expected to self-heal, so OutOfSync/Missing is a real failure rather
-// than an operator who simply hasn't synced a manual app yet.
-func argoIsAutomated(app *unstructured.Unstructured) bool {
-	automated, found, _ := unstructured.NestedMap(app.Object, "spec", "syncPolicy", "automated")
-	if !found {
-		return false
-	}
-	// Newer Argo CD can disable auto-sync without removing the block, via
-	// spec.syncPolicy.automated.enabled: false — treat that as manual so an
-	// intentionally-unsynced app isn't flagged for OutOfSync/Missing.
-	if enabled, ok, _ := unstructured.NestedBool(automated, "enabled"); ok && !enabled {
-		return false
-	}
-	return true
-}
-
-// argoErrorCondition returns the first status.conditions entry whose type names
-// an error (ComparisonError / InvalidSpecError / SyncError). Argo writes these
-// as {type, message} without a status field, so FindFalseCondition can't match
-// them.
-func argoErrorCondition(app *unstructured.Unstructured) (condType, message string, found bool) {
-	conds, ok, _ := unstructured.NestedSlice(app.Object, "status", "conditions")
-	if !ok {
-		return "", "", false
-	}
-	for _, c := range conds {
-		cm, ok := c.(map[string]any)
-		if !ok {
-			continue
-		}
-		ct, _ := cm["type"].(string)
-		if strings.Contains(ct, "Error") {
-			msg, _ := cm["message"].(string)
-			return ct, msg, true
-		}
-	}
-	return "", "", false
-}
-
-// detectFluxProblems flags Flux Kustomizations/HelmReleases whose Ready condition
-// is False for a genuine (non-in-progress) reason. Unlike the shared
-// packages.IsTransientConditionReason set used for health display, this uses a
-// NARROW in-progress set so genuinely-stuck states it treats as transient
-// (ArtifactFailed, ChartNotReady) DO surface as issues. Skips suspended objects
-// and stale-generation conditions (controller hasn't observed the current spec).
-func detectFluxProblems(items []*unstructured.Unstructured, kind, group string, now time.Time) []Detection {
-	var out []Detection
-	for _, obj := range items {
-		if suspend, ok, _ := unstructured.NestedBool(obj.Object, "spec", "suspend"); ok && suspend {
-			continue
-		}
-		_, reason, msg, since, ok := conditions.FindFalseCondition(obj, "Ready")
-		if !ok || conditions.IsInProgressForIssues(reason) {
-			continue
-		}
-		// status.conditions stale relative to spec → mid-reconcile, not failed.
-		if gen := obj.GetGeneration(); gen > 0 {
-			if observed, ok, _ := unstructured.NestedInt64(obj.Object, "status", "observedGeneration"); ok && observed > 0 && observed < gen {
-				continue
-			}
-		}
-		age := now.Sub(obj.GetCreationTimestamp().Time)
-		d := since
-		if d == 0 {
-			d = age
-		}
-		displayReason := reason
-		if displayReason == "" {
-			displayReason = "Ready=False"
-		}
-		p := gitopsProblem(kind, group, obj.GetNamespace(), obj.GetName(), "high", displayReason, msg, age)
-		p.DurationSeconds = int64(d.Seconds())
-		p.Duration = FormatAge(d)
-		out = append(out, p)
-	}
-	return out
-}
