+++ internal/k8s/detection_test.go
